build: clarify usage of stable versions
Update the documentation to use a more recent version in the examples.
With that version people shouldn't face the issue that we've seen in
2020 when it came to Python2 deprecation, pycryptodome and repo updates
where symlink creation had been changed.

Also add an extra clarification that we recommend to always use the most
recent stable commit if there is no other strong reason to not do so.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>