coding standard: defining unsigned integer constants using U()
Misra C rule 7.2 says:
"A 'u' or 'U' suffix shall be applied to all integer constants that are
represented in an unsigned type"

Besides Misra this makes sense anyway to make the sign and size of the
integer well defined instead of depending on the size of the constant or
version of C standard a compiler is using.

Do this with the U() macro instead of an explicit 'U' at the end to make
the defines usable from assembly too.

Acked-by: Jerome Forissier <jerome@forissier.org>
Signed-off-by: Jens Wiklander <jens.wiklander@linaro.org>