stm32mp1: remove sudo command when loading SDcard image
Remove the sudo command from the shell instruction . When things
can be made without sudo admin rights, it is really preferred to
use standard rights hence not recommending this in this documentation.

The fact is it may not be necessary to use sudo right to update
an SDcard device on a host, depending on the host configuration.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>