crypto: document the NXP SE05X libseetec interface
Even though the secure element is expected to be controlled from the
PKCS#11 TA, the driver also provides an interface to the OP-TEE client.

Using this interface, a priviledged application can access all the
functionality in the device by encoding the APDU frames in the normal
world and sending them to the device.

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>