building: qemu: update fetch/build/run command example
The repo command fetches code in the current directory so it is best to
run it inside a directory created specifically for that purpose.

The mentions about disabling ASLR are not exact (TA ASLR has its own
config flag) and are not very relevant to the short descriptions so
remove them.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Etienne Carriere <etienne.carriere@linaro.org>
Reviewed-by: Joakim Bech <joakim.bech@linaro.org>