rockpi4 documentation fixes
building/devices/rockpi4.rst: The "RockPi4" wiki
page has been removed and combined with "Rock4" so
update our external link so it still works.

building/gits/build.rst: It looks like the commit that
added rockpi4 build documentation added Rock Pi 4 to the list/table
in "Platforms Supported" but did not add it the table in the
"Current Version" section. This PR adds it to the second table.

Signed-off-by: Sean Gilligan <sean@msgilligan.com>