architecutre: porting: describe tee.bin and tee-raw.bin
For someone just trying to launch OP-TEE and print "Hello World!",
tee.bin seems like the obvious binary to use. However it contains a
non-executable header(?!) which can cause some lost days of debugging.

Add a mention that tee-raw.bin should be used as it can just be jumped
to.

Signed-off-by: Caleb Connolly <caleb.connolly@linaro.org>