building: add Fedora 42 support
Add a Dockerfile for Fedora 42, which is based on the Dockerfile already
existing for Ubuntu. Fedora uses different names of packages, hence the
list looks a bit different. We've also added an editor (neovim), to have
something powerful ready to be used when going into a container and
running commands.

Signed-off-by: Joakim Bech <joakim.bech@gmail.com>
Tested-by: Joakim Bech <joakim.bech@gmail.com>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>